✨ Refactor dependencies: move director-related dependencies to a dedicated _dependencies module
