+++ services/catalog/src/simcore_service_catalog/api/_dependencies/user_groups.py
