+++ services/catalog/src/simcore_service_catalog/api/rest/_services.py
@@ -24,9 +24,9 @@ from ...db.repositories.groups import GroupsRepository
 from ...db.repositories.services import ServicesRepository
 from ...models.services_db import ServiceAccessRightsAtDB, ServiceMetaDataDBGet
 from ...services.director import DirectorApi
-from ..dependencies.database import get_repository
-from ..dependencies.director import get_director_api
-from ..dependencies.services import get_service_from_manifest
+from .._dependencies.database import get_repository
+from .._dependencies.director import get_director_api
+from .._dependencies.services import get_service_from_manifest
 
 _logger = logging.getLogger(__name__)
