+++ services/catalog/src/simcore_service_catalog/api/rest/_services_extras.py
@@ -5,7 +5,7 @@ from models_library.api_schemas_directorv2.services import ServiceExtras
 from models_library.services import ServiceKey, ServiceVersion
 
 from ...services.director import DirectorApi
-from ..dependencies.director import get_director_api
+from .._dependencies.director import get_director_api
 
 router = APIRouter()
