+++ services/catalog/src/simcore_service_catalog/api/rest/_services_labels.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends
 from models_library.services import ServiceKey, ServiceVersion
 
 from ...services.director import DirectorApi
-from ..dependencies.director import get_director_api
+from .._dependencies.director import get_director_api
 
 router = APIRouter()
