+++ services/catalog/src/simcore_service_catalog/api/rest/_services_specifications.py
@@ -13,8 +13,8 @@ from ..._constants import RESPONSE_MODEL_POLICY
 from ...db.repositories.groups import GroupsRepository
 from ...db.repositories.services import ServicesRepository
 from ...services.function_services import is_function_service
-from ..dependencies.database import get_repository
-from ..dependencies.services import get_default_service_specifications
+from .._dependencies.database import get_repository
+from .._dependencies.services import get_default_service_specifications
 
 router = APIRouter()
 _logger = logging.getLogger(__name__)
