+++ services/catalog/tests/unit/test_services_director.py
@@ -15,7 +15,7 @@ from models_library.services_metadata_published import ServiceMetaDataPublished
 from pytest_simcore.helpers.monkeypatch_envs import setenvs_from_dict
 from pytest_simcore.helpers.typing_env import EnvVarsDict
 from respx.router import MockRouter
-from simcore_service_catalog.api.dependencies.director import get_director_api
+from simcore_service_catalog.api._dependencies.director import get_director_api
 from simcore_service_catalog.services.director import DirectorApi
