+++ services/catalog/tests/unit/with_dbs/test_services_services_api.py
@@ -16,7 +16,7 @@ from models_library.users import UserID
 from pydantic import TypeAdapter
 from pytest_simcore.helpers.catalog_services import CreateFakeServiceDataCallable
 from respx.router import MockRouter
-from simcore_service_catalog.api.dependencies.director import get_director_api
+from simcore_service_catalog.api._dependencies.director import get_director_api
 from simcore_service_catalog.db.repositories.groups import GroupsRepository
 from simcore_service_catalog.db.repositories.services import ServicesRepository
 from simcore_service_catalog.services import manifest, services_api
